pyproject.toml
[project]
name = "rentvsbuyai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "fuzzywuzzy>=0.18.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.5",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.1",
    "langsmith>=0.4.4",
    "mcp-use>=1.3.3",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "python-mcp-client>=0.1.2",
    "requests>=2.32.4",
    "tabulate>=0.9.0",
    "typing-extensions>=4.14.0",
]
